+++ guide/src/news.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## main
+
+- new `export` command to write the password database as a JSON file
+
 ## 25.8
 
 - new `gc` command to rebuild the database file, repacking it into a minimal amount of disk space